moved config to config folder
